Roles and permissions api refactored

+++ app/Http/Controllers/API/V1/RoleAPIController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers\API\V1;
 
+use App\Http\Controllers\AppBaseController;
 use App\Http\Requests\API\CreateRoleAPIRequest;
 use App\Http\Requests\API\UpdateRoleAPIRequest;
 use App\Models\Role;
 use App\Repositories\RoleRepository;
 use Illuminate\Http\Request;
-use App\Http\Controllers\AppBaseController;
 use Response;
+use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 /**
  * Class RoleAPIController
@@ -39,24 +40,25 @@ class RoleAPIController extends AppBaseController
             $request->get('limit')
         );
 
-        return $this->sendResponse($roles->toArray(), 'Roles retrieved successfully');
+        return $this->sendResponse($roles->toArray(), 'Roles retrieved successfully.');
     }
 
     /**
      * Store a newly created Role in storage.
      * POST /roles
-     *
      * @param CreateRoleAPIRequest $request
      *
-     * @return Response
+     * @throws \App\Exceptions\ApiOperationFailedException
+     *
+     * @return \Illuminate\Http\JsonResponse
      */
     public function store(CreateRoleAPIRequest $request)
     {
         $input = $request->all();
 
-        $role = $this->roleRepository->create($input);
+        $role = $this->roleRepository->store($input);
 
-        return $this->sendResponse($role->toArray(), 'Role saved successfully');
+        return $this->sendResponse($role->toArray(), 'Role saved successfully.');
     }
 
     /**
@@ -70,19 +72,20 @@ class RoleAPIController extends AppBaseController
     public function show($id)
     {
         /** @var Role $role */
-        $role = $this->roleRepository->findOrFail($id);
+        $role = $this->roleRepository->find($id);
 
-        return $this->sendResponse($role->toArray(), 'Role retrieved successfully');
+        return $this->sendResponse($role->toArray(), 'Role retrieved successfully.');
     }
 
     /**
      * Update the specified Role in storage.
      * PUT/PATCH /roles/{id}
-     *
      * @param int $id
      * @param UpdateRoleAPIRequest $request
      *
-     * @return Response
+     * @throws \App\Exceptions\ApiOperationFailedException
+     *
+     * @return \Illuminate\Http\JsonResponse
      */
     public function update($id, UpdateRoleAPIRequest $request)
     {
@@ -93,7 +96,7 @@ class RoleAPIController extends AppBaseController
 
         $role = $this->roleRepository->update($input, $id);
 
-        return $this->sendResponse($role->toArray(), 'Role updated successfully');
+        return $this->sendResponse($role->toArray(), 'Role updated successfully.');
     }
 
     /**
@@ -109,10 +112,16 @@ class RoleAPIController extends AppBaseController
     public function destroy($id)
     {
         /** @var Role $role */
-        $role = $this->roleRepository->findOrFail($id);
+        $role = Role::withCount('perms')->findOrFail($id);
+        if ($role->perms_count > 0) {
+            return $this->sendError(
+                'Role is attached with one or more permissions.',
+                HttpResponse::HTTP_UNPROCESSABLE_ENTITY
+            );
+        }
 
         $role->delete();
 
-        return $this->sendResponse($id, 'Role deleted successfully');
+        return $this->sendResponse($id, 'Role deleted successfully.');
     }
 }
